requirements.txt
asgiref==3.3.1
coverage==5.3
djangorestframework==3.12.2
pytz==2020.4
sqlparse==0.4.1
gunicorn==20.0.4
Django>=3.0,<4.0
psycopg2-binary>=2.8